Refactoring.
This way I try to have more randomization.
I've observed that on certain conditions Random might produce first result very similar independently of seed.
For example when in a loop new Random() is created for consequent seed values (1, 2, 3...) and then nextInt(32) is called - the value will be the same.
That also works for nextBoolean()